@@ -1,17 +1,25 @@
 import React from 'react'
 import {
   BrowserRouter as Router,
-  Route
+  Route,
+  Routes
 } from 'react-router-dom'
 
+import PageLayout from './components/common/PageLayout'
 import Home from './components/Home'
 import Quiz from './components/Quiz/Index'
+import Result from './components/Result/Index'
 
 const App = () => {
   return (
     <Router>
-      <Route path="/" exact component={Home} />
-      <Route path="/quiz" exact component={Quiz} />
+      <PageLayout>
+        <Routes>
+          <Route path="/" element={<Home />} />
+          <Route path="/quiz" element={<Quiz />} />
+          <Route path="/result" element={<Result />} />
+        </Routes>
+      </PageLayout>
     </Router>
   )
 }
